tests/requirements.txt
jinja2==3.0.3
MarkupSafe==2.1.1
pyyaml==6.0
pytest==4.6.5
jsonschema==4.5.1
